clean up test class
