+++ src/test/java/posmy/interview/boot/service/MemberDeleteServiceTest.java
@@ -3,14 +3,9 @@ package posmy.interview.boot.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.dao.DataAccessResourceFailureException;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import posmy.interview.boot.entity.MyUser;
 import posmy.interview.boot.repos.MyUserRepository;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -20,15 +15,9 @@ import static org.mockito.Mockito.*;
 class MemberDeleteServiceTest {
 
     private final MyUserRepository myUserRepository = mock(MyUserRepository.class);
-    private final PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 
     private final MemberDeleteService memberDeleteService =
             new MemberDeleteService(myUserRepository);
-
-    @Captor
-    private final ArgumentCaptor<MyUser> userCaptor = ArgumentCaptor.forClass(MyUser.class);
-
-    private MyUser existingUser;
 
     @BeforeEach
     void setup() {
